+++ CHANGELOG.md
@@ -3,6 +3,9 @@
   * Event:
     * Removal of add\_field method (use add\_column instead)
     * Removal of fields attribute (use columns instead)
+  * Section:
+    * subscription\_level attribute is now a Fixnum not Symbol
+    * Addition of subscription\_level\_name method to get the name of the subscription level for the section
 
 ## Version 0.1.17